fix(schema): resolve multiple Hibernate validation mismatches
- Payment.date: change LocalDate to LocalDateTime, add @Column(name="payment_date")
- StockMovement.date: change LocalDate to LocalDateTime (DB has TIMESTAMP)
- StockMovementDto.date: change LocalDate to LocalDateTime to match entity
- CreateStockMovementRequest: convert LocalDate to LocalDateTime in service
- PurchaseOrderLine.receivedQty: change Integer to BigDecimal (DB has NUMERIC)
- PurchaseOrderLineDto.receivedQty: change Integer to BigDecimal
- PurchaseOrderService: update receivedQty usage to BigDecimal arithmetic
- HelpdeskTeam.defaultPriority: change length from 10 to 1 (DB has VARCHAR(1))
- PaymentDto.date: change LocalDate to LocalDateTime to match entity
- InvoiceService: remove unnecessary toLocalDate() conversion

+++ src/main/java/com/erp/purchasing/dto/PurchaseOrderLineDto.java
@@ -20,6 +20,6 @@ public class PurchaseOrderLineDto {
     private BigDecimal unitPrice;
     private BigDecimal discount;
     private BigDecimal lineTotal;
-    private Integer receivedQty;
+    private BigDecimal receivedQty;
     private String notes;
 }
